+++ R/reprex.R
@@ -7,19 +7,18 @@
 #' be
 #'
 #' * read from clipboard
-#' * read from current selection or active document in RStudio
-#'   (with [reprex_addin()])
 #' * provided directly as expression, character vector, or string
 #' * read from file
+#' * read from current selection or active document in RStudio
 #'
 #' reprex can also be used for syntax highlighting (with or without rendering);
 #' see below for more.
 #'
 #' @section Details:
-#' The usual "code + commented output" is returned invisibly, put on the
-#' clipboard, and written to file. An HTML preview displays in RStudio's Viewer
-#' pane, if available, or in the default browser, otherwise. Leading `"> "`
-#' prompts, are stripped from the input code. Read more at
+#' The usual "code + commented output" is returned invisibly, possibly put on
+#' the clipboard, and written to file. An HTML preview displays in RStudio's
+#' Viewer pane, if available, or in the default browser, otherwise. Leading
+#' `"> "` prompts, are stripped from the input code. Read more at
 #' <https://reprex.tidyverse.org/>.
 #'
 #' reprex sets specific [knitr options](https://yihui.org/knitr/options/):
@@ -28,7 +27,7 @@
 #'   reporting.
 #' * reprex also sets knitr's `upload.fun`. It defaults to
 #'   [knitr::imgur_upload()] so figures produced by the reprex appear properly
-#'   on GitHub, Stack Overflow, or Discourse. Note that `imgur_upload()`
+#'   on GitHub, Stack Overflow, Discourse, and Slack. Note that `imgur_upload()`
 #'   requires the packages httr and xml2. When `venue = "r"`, `upload.fun` is
 #'   set to `identity()`, so that figures remain local. In that case, you may
 #'   also want to set `outfile`.
@@ -49,7 +48,11 @@
 #' article](https://reprex.tidyverse.org/articles/articles/rtf.html)
 #'
 #' @param x An expression. If not given, `reprex()` looks for code in
-#'   `input` or on the clipboard, in that order.
+#'   `input`. If `input` is not provided, `reprex()` looks on the clipboard.
+#'
+#'   When the clipboard is structurally unavailable, i.e. on RStudio Server or
+#'   RStudio Cloud, `reprex()` consults the current selection instead of the
+#'   clipboard.
 #' @param input Character. If has length one and lacks a terminating newline,
 #'   interpreted as the path to a file containing reprex code. Otherwise,
 #'   assumed to hold reprex code as character vector.