+++ tests/testthat/helper.R
@@ -2,6 +2,10 @@ expect_error_free <- function(...) {
   expect_error(..., regexp = NA)
 }
 
+with_mock <- function(..., .parent = parent.frame()) {
+  mockr::with_mock(..., .parent = .parent, .env = "reprex")
+}
+
 # 1. creates a subdirectory within session temp
 # 2. makes that the current working directory
 # 3. schedules these cleanup actions for when env goes out of scope: