+++ tests/testthat/test-utils.R
@@ -1,5 +1,18 @@
 test_that("locate_input() works", {
-  expect_identical("clipboard", locate_input(NULL))
+  with_mock(
+    clipboard_available = function() TRUE,
+    expect_identical("clipboard", locate_input(NULL))
+  )
+  with_mock(
+    clipboard_available = function() FALSE,
+    in_rstudio = function() TRUE,
+    expect_identical("selection", locate_input(NULL))
+  )
+  with_mock(
+    clipboard_available = function() FALSE,
+    in_rstudio = function() FALSE,
+    expect_null(locate_input(NULL))
+  )
   expect_identical("path", locate_input(path_temp()))
   expect_identical("input", locate_input(c("a", "b")))
   expect_identical("input", locate_input("a\n"))